Installed styled components and began with the re-branding.
