+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "editor.tabSize": 2,
+    "editor.formatOnSave": true
+}
