--- index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<meta http-equiv="X-UA-Compatible" content="ie=edge">
-	<link rel="stylesheet" href="./dist/css/portfolio.717ebadd7270ef759573.css">
-	<link rel="shortcut icon" href="./src/images/resume.png" type="image/x-icon">
-	<title>Portfolio</title>
-</head>
-
-<body>
-	<div id="app"></div>
-	<script src="./dist/js/portfolio.717ebadd7270ef759573.js"></script>
-</body>                               
-</html>
