+++ package.json
@@ -4,8 +4,8 @@
   "description": "Portfolio where I show my professional carreer.",
   "main": "index.js",
   "scripts": {
-    "build": "webpack -p",
-    "build:dev": "webpack-dev-server --mode development --config ./webpack.dev.config.js"
+    "build": "webpack",
+    "dev": "webpack-dev-server"
   },
   "repository": {
     "type": "git"
@@ -13,34 +13,33 @@
   "author": "@jfbenavid",
   "license": "MIT",
   "devDependencies": {
-    "babel-core": "6.26.3",
-    "babel-loader": "7.1.5",
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-preset-env": "1.7.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "6.24.1",
-    "babel-preset-stage-2": "6.24.1",
-    "clean-webpack-plugin": "0.1.19",
-    "css-loader": "1.0.0",
-    "file-loader": "1.1.11",
-    "font-awesome": "^4.7.0",
-    "mini-css-extract-plugin": "0.4.1",
-    "nib": "1.1.2",
-    "optimize-css-assets-webpack-plugin": "^5.0.1",
-    "rupture": "0.7.1",
-    "style-loader": "0.22.1",
-    "stylus": "0.54.5",
-    "stylus-loader": "3.0.2",
-    "url-loader": "1.0.1",
-    "webpack": "4.16.5",
-    "webpack-cli": "3.1.0",
-    "webpack-dev-server": "^3.7.1"
+    "@babel/core": "^7.5.5",
+    "@babel/preset-env": "^7.5.5",
+    "@babel/preset-react": "^7.0.0",
+    "babel-eslint": "^10.0.2",
+    "babel-loader": "^8.0.6",
+    "eslint": "^6.2.1",
+    "html-webpack-plugin": "^3.2.0",
+    "standard": "^14.0.2",
+    "webpack": "^4.39.2",
+    "webpack-cli": "^3.3.7",
+    "webpack-dev-server": "^3.8.0"
   },
   "dependencies": {
-    "react": "16.4.2",
-    "react-dom": "16.4.2",
-    "react-fontawesome": "^1.6.1",
-    "react-redux": "^5.0.7",
-    "redux": "^4.0.0"
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
+    "react-redux": "^5.1.2",
+    "redux": "^4.0.5",
+    "styled-components": "^5.0.1"
+  },
+  "standard": {
+    "ignore": [],
+    "parser": "babel-eslint"
+  },
+  "eslintConfig": {
+    "extends": [
+      "./node_modules/standard/eslintrc.json"
+    ],
+    "parser": "babel-eslint"
   }
 }
