--- src/js/portfolio/header/containers/Header.js
@@ -1,14 +1,0 @@
-import React, { Component } from 'react';
-import '../styles/header.styl';
-
-export default class Header extends Component {
-    render() {
-        return (
-            <header className="header">
-                <div className="container hero">
-                    {this.props.children}
-                </div>
-            </header>
-        );
-    }
-}
