+++ src/portfolio/styles.styl
@@ -1,4 +1,3 @@
-@import '~/font-awesome/css/font-awesome.min.css'
 @import url('https://fonts.googleapis.com/css?family=Montserrat:300,400,500,600,700|Roboto:300,400,500,700')
 linesColor = #3B3B3B
 testColor = #222222
