+++ src/styles/theme-provider.js
@@ -1,0 +1,5 @@
+export const theme = {
+    colors: {
+        grey: '#3B3B3B'
+    }
+}
